moved uri parser to autware_test_utils helper
