+++ common/autoware_test_utils/CMakeLists.txt
@@ -5,16 +5,13 @@ find_package(autoware_cmake REQUIRED)
 autoware_package()
 
 ament_auto_add_library(autoware_test_utils SHARED
-  src/autoware_test_utils.cpp)
-
-ament_auto_add_library(mock_data_parser SHARED
-  src/mock_data_parser.cpp)
+  src/autoware_test_utils.cpp
+  src/mock_data_parser.cpp
+)
 
 ament_auto_add_executable(topic_snapshot_saver src/topic_snapshot_saver.cpp)
 
-target_link_libraries(mock_data_parser
-  yaml-cpp
-)
+target_link_libraries(topic_snapshot_saver autoware_test_utils)
 
 if(BUILD_TESTING)
   ament_auto_add_gtest(test_autoware_test_utils
