+++ common/autoware_test_utils/src/autoware_test_utils.cpp
@@ -312,4 +312,18 @@ PathWithLaneId loadPathWithLaneIdInYaml()
 
   return parse<PathWithLaneId>(yaml_path);
 }
+
+std::optional<std::string> resolve_pkg_share_uri(const std::string & uri_path)
+{
+  std::smatch match;
+  std::regex pattern(R"(package://([^/]+)/(.+))");
+  if (std::regex_match(uri_path, match, pattern)) {
+    const std::string pkg_name = ament_index_cpp::get_package_share_directory(match[1].str());
+    const std::string resource_path = match[2].str();
+    const auto path = std::filesystem::path(pkg_name) / std::filesystem::path(resource_path);
+    return std::filesystem::exists(path) ? std::make_optional<std::string>(path) : std::nullopt;
+  }
+  return std::nullopt;
+}
+
 }  // namespace autoware::test_utils
